Split `test_geo.py` into processor-specific files
